pgsql transaction logic updated to v1.9.0
Major changes to updated the PostgreSQL related code to v1.9.0.

Debugging code was been completed and additional code added to help
validate the transaction logic.  Updated the formatting look of the
outpout to more easily read what is going on.

Additional checks are made to verify the number of the items returned in
column array match what the database transactions says there are.
Verbosity of individual checks still vary.  There are certain elements
that the spec does not guarantee the number of elements for.  These
particular pieces are not checked and are marked FIXME in case this
interpretation is incorrect.

Fixed string tokenizer to correctly parse string.  Also added another
tokenizer to parse arrays out of arrays.  Should probably be combined at
some point.  There are still potential strings that will break the
tokenizer, but it is not expected that EGen will create those types of
strings.  Exceptions are noted in the code.  Move the tokenizer code
from the TxnBasedB.h header file to the transaction.h header file.  It
is supposed to be helper inlined functions as opposed to a class method.

Several bugs fixed the PostgreSQL stored functions.  All stored
functions should be updated to spec, except for the Market Feed, which
is noted in the code.  For the PostgreSQL stored functions (not the
pl/psql code):

* Fixed various data size problems when various id's were set to int.
  All have been fixed to be long's
* Fixed a incorrectly sized news variable in Security Detail frame 1.
* Properly return day data in Security Detail frame 1.
* Fixed a bug in Trade Lookup frame 1 when checking is_cash.
* Properly return trade history data in Trade Lookup Frame 1.
* Fixed the handling of trade history data in Trade Update frame 1.
* Trade update frame 3 was not escaping a string correctly for one of
  its queries.
* Fixed queries in the Trade Result frame to handle escaped strings, and
  looping through return values.
* Added error check for the Data Maintenance frame in the case where the
  table name doesn't match.

The driver code has been updated to match the updated stored functions.
The order of reading in the values from the stored functions are mostly
ordered by input/output order of the transaction descriptions.  The
intent is to be able to easily proof things against the specification.

* Use strncpy over strcpy to prevent possible memory violations cause by
  bad data returned by the database.
* Correctly parse history output in the Customer Position frame 2.
* More correctly implement the Market Feed frame.  Limitations in the
  ability to use transactions in PostgreSQL stored function limits the
  implementation for that database.
* Correctly parse the fin and dts data in the Security Detail frame 1.
* Correctly parse trade history data in all of the Trade Lookup frames.
* Fixed memory violation bug caused by misuse of sscanf in the Trade
  Update frames when reading timestamps.
* The correct isolation level was not set to REPEATABLE READ for all of
  the possible Trade Update frame transaction starting locations.
* Fixed MarketFeed to return proper status code on error.
* Fixed call to Security Detail frame 1 to use a DATE instead of a
  TIMESTAMP datatype.
* Removed COMMIT call in Trade Result frame 5 that should not be there.

+++ code/storedproc/pgsql/c/trade_order.c
@@ -37,7 +37,7 @@
 #define TOF2_1 \
 		"SELECT ap_acl\n" \
 		"FROM account_permission\n" \
-		"WHERE ap_ca_id = %d\n" \
+		"WHERE ap_ca_id = %ld\n" \
 		"  AND ap_f_name = '%s'\n" \
 		"  AND ap_l_name = '%s'\n" \
 		"  AND ap_tax_id = '%s'"
@@ -151,11 +151,11 @@ PG_MODULE_MAGIC;
 #endif
 
 /* Prototypes. */
-void dump_tof1_inputs(int);
-void dump_tof2_inputs(int, char *, char *, char *);
-void dump_tof3_inputs(int, int, int, int, char *, char *, char *, int, int,
+void dump_tof1_inputs(long);
+void dump_tof2_inputs(long, char *, char *, char *);
+void dump_tof3_inputs(long, long, int, int, char *, char *, char *, int, int,
 		char *, int, char *, double, char *);
-void dump_tof4_inputs(int, int, double, char *, int, int, double, char *,
+void dump_tof4_inputs(long, long, double, char *, int, int, double, char *,
 		char *, int, char *, int);
 
 Datum TradeOrderFrame1(PG_FUNCTION_ARGS);
@@ -168,32 +168,32 @@ PG_FUNCTION_INFO_V1(TradeOrderFrame2);
 PG_FUNCTION_INFO_V1(TradeOrderFrame3);
 PG_FUNCTION_INFO_V1(TradeOrderFrame4);
 
-void dump_tof1_inputs(int acct_id)
+void dump_tof1_inputs(long acct_id)
 {
 	elog(NOTICE, "TOF1: INPUTS START");
-	elog(NOTICE, "TOF1: acct_id %d", acct_id);
+	elog(NOTICE, "TOF1: acct_id %ld", acct_id);
 	elog(NOTICE, "TOF1: INPUTS END");
 }
 
-void dump_tof2_inputs(int acct_id, char *exec_f_name, char *exec_l_name,
+void dump_tof2_inputs(long acct_id, char *exec_f_name, char *exec_l_name,
 		char *exec_tax_id)
 {
 	elog(NOTICE, "TOF2: INPUTS START");
-	elog(NOTICE, "TOF2: acct_id %d", acct_id);
+	elog(NOTICE, "TOF2: acct_id %ld", acct_id);
 	elog(NOTICE, "TOF2: exec_f_name %s", exec_f_name);
 	elog(NOTICE, "TOF2: exec_l_name %s", exec_l_name);
 	elog(NOTICE, "TOF2: exec_tax_id %s", exec_tax_id);
 	elog(NOTICE, "TOF2: INPUTS END");
 }
 
-void dump_tof3_inputs(int acct_id, int cust_id, int cust_tier, int is_lifo,
+void dump_tof3_inputs(long acct_id, long cust_id, int cust_tier, int is_lifo,
 		char *issue, char *st_pending_id, char *st_submitted_id,
 		int tax_status, int trade_qty, char *trade_type_id, int type_is_margin,
 		char *co_name, double requested_price, char *symbol)
 {
 	elog(NOTICE, "TOF3: INPUTS START");
-	elog(NOTICE, "TOF3: acct_id %d", acct_id);
-	elog(NOTICE, "TOF3: cust_id %d", cust_id);
+	elog(NOTICE, "TOF3: acct_id %ld", acct_id);
+	elog(NOTICE, "TOF3: cust_id %ld", cust_id);
 	elog(NOTICE, "TOF3: cust_tier %d", cust_tier);
 	elog(NOTICE, "TOF3: is_lifo %d", is_lifo);
 	elog(NOTICE, "TOF3: issue %s", issue);
@@ -209,13 +209,13 @@ void dump_tof3_inputs(int acct_id, int cust_id, int cust_tier, int is_lifo,
 	elog(NOTICE, "TOF3: INPUTS END");
 }
 
-void dump_tof4_inputs(int acct_id, int broker_id, double charge_amount,
+void dump_tof4_inputs(long acct_id, long broker_id, double charge_amount,
 		char *exec_name, int is_cash, int is_lifo, double requested_price,
 		char *status_id, char *symbol, int trade_qty, char *trade_type_id,
 		int type_is_market) {
 	elog(NOTICE, "TOF4: INPUTS START");
-	elog(NOTICE, "TOF4: acct_id %d", acct_id);
-	elog(NOTICE, "TOF4: broker_id %d", broker_id);
+	elog(NOTICE, "TOF4: acct_id %ld", acct_id);
+	elog(NOTICE, "TOF4: broker_id %ld", broker_id);
 	elog(NOTICE, "TOF4: charge_amount %8.2f", charge_amount);
 	elog(NOTICE, "TOF4: exec_name %s", exec_name);
 	elog(NOTICE, "TOF4: is_cash %d", is_cash);
@@ -412,7 +412,7 @@ Datum TradeOrderFrame2(PG_FUNCTION_ARGS)
 
 		enum tof2 { i_ap_acl=0, i_status };
 
-		int acct_id = PG_GETARG_INT64(0);
+		long acct_id = PG_GETARG_INT64(0);
 		char *exec_f_name_p = (char *) PG_GETARG_TEXT_P(1);
 		char *exec_l_name_p = (char *) PG_GETARG_TEXT_P(2);
 		char *exec_tax_id_p = (char *) PG_GETARG_TEXT_P(3);
